# Conflicts:
#	crates/sori-core/src/text_injection.rs

+++ crates/sori-core/src/text_injection.rs
@@ -541,10 +541,10 @@ pub mod windows {
                 },
             };
             let inputs = [
-                key(VK_CONTROL, 0),
-                key(VK_V, 0),
-                key(VK_V, KEYEVENTF_KEYUP),
-                key(VK_CONTROL, KEYEVENTF_KEYUP),
+                key(VK_CONTROL as u16, 0),
+                key(VK_V as u16, 0),
+                key(VK_V as u16, KEYEVENTF_KEYUP),
+                key(VK_CONTROL as u16, KEYEVENTF_KEYUP),
             ];
             let sent = unsafe {
                 SendInput(
@@ -576,7 +576,12 @@ pub mod windows {
                 INPUT, INPUT_0, INPUT_KEYBOARD, KEYBDINPUT, KEYEVENTF_KEYUP, SendInput, VK_CONTROL,
                 VK_LMENU, VK_LSHIFT, VK_LWIN,
             };
-            let keys = [VK_CONTROL, VK_LSHIFT, VK_LMENU, VK_LWIN];
+            let keys = [
+                VK_CONTROL as u16,
+                VK_LSHIFT as u16,
+                VK_LMENU as u16,
+                VK_LWIN as u16,
+            ];
             let inputs: Vec<INPUT> = keys
                 .into_iter()
                 .map(|vk| INPUT {
